add context processor to retrieve user login information
By defining a custom context processor on the settings file, one is able
to use the variable 'logged_in' on the templates, having it show either
True for logged in users or False otherwise. Django context processors
are but mere functions defined by the user, and the ones already
provided by Django itself.